refactor: UploadFileForm component
